requirements/prod.txt
-r base.txt

psycopg2-binary
dj-database-url
django-compressor
gunicorn
whitenoise
